@@ -3,24 +3,24 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@aws-sdk/client-cognito-identity": "^3.42.0",
-    "@aws-sdk/client-dynamodb": "^3.42.0",
-    "@aws-sdk/credential-provider-cognito-identity": "^3.42.0",
-    "@aws-sdk/credential-providers": "^3.42.0",
-    "@aws-sdk/util-dynamodb": "^3.42.0",
-    "@emotion/react": "^11.5.0",
-    "@emotion/styled": "^11.3.0",
-    "@mui/icons-material": "^5.1.0",
-    "@mui/material": "^5.1.0",
-    "@reduxjs/toolkit": "^1.6.2",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "@types/jest": "^27.0.2",
-    "@types/node": "^16.11.7",
-    "@types/react": "^17.0.34",
+    "@aws-sdk/client-cognito-identity": "^3.45.0",
+    "@aws-sdk/client-dynamodb": "^3.45.0",
+    "@aws-sdk/credential-provider-cognito-identity": "^3.45.0",
+    "@aws-sdk/credential-providers": "^3.45.0",
+    "@aws-sdk/util-dynamodb": "^3.45.0",
+    "@emotion/react": "^11.7.1",
+    "@emotion/styled": "^11.6.0",
+    "@mui/icons-material": "^5.2.5",
+    "@mui/material": "^5.2.5",
+    "@reduxjs/toolkit": "^1.7.1",
+    "@testing-library/jest-dom": "^5.16.1",
+    "@testing-library/react": "^12.1.2",
+    "@testing-library/user-event": "^13.5.0",
+    "@types/jest": "^27.0.3",
+    "@types/node": "^17.0.4",
+    "@types/react": "^17.0.38",
     "@types/react-dom": "^17.0.11",
-    "@types/react-redux": "^7.1.20",
+    "@types/react-redux": "^7.1.21",
     "@types/react-router": "^5.1.17",
     "@types/react-router-dom": "^5.3.2",
     "amazon-cognito-identity-js": "^5.2.3",
@@ -29,11 +29,11 @@
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-redux": "^7.2.6",
-    "react-router-dom": "^5.3.0",
-    "react-scripts": "4.0.3",
+    "react-router-dom": "^6.2.1",
+    "react-scripts": "^5.0.0",
     "redux": "^4.1.2",
-    "typescript": "^4.4.4",
-    "web-vitals": "^1.0.1"
+    "typescript": "^4.5.4",
+    "web-vitals": "^2.1.2"
   },
   "scripts": {
     "start": "react-scripts start",
