@@ -1,2 +1,3 @@
 import type { RootState } from '../store';
-export const selectMyProfile = (state: RootState) => state.profile.myProfile;
+import { ProfileDb } from './types';
+export const selectMyProfile = (state: RootState) : ProfileDb | null => state.profile.myProfile || null; 
