@@ -3,4 +3,4 @@ import { values } from 'lodash';
 import { ProgramDb } from './types';
 
 export const selectAllPrograms = (state : RootState) : ProgramDb[] => values(state.allPrograms);
-export const selectProgram = (programId : string) => (state : RootState) => state.allPrograms[programId];
+export const selectProgram = (programId? : string|null) => (state : RootState) : ProgramDb|null => programId ? state.allPrograms[programId] : null;
