@@ -1,9 +1,11 @@
-import { SetStateAction, useState } from 'react';
-import { Grid, FormControl, InputLabel, MenuItem, Select, Typography } from '@mui/material';
+import React, { SetStateAction, useState } from 'react';
+import { Grid, FormControl, InputLabel, MenuItem, Select, Container, CssBaseline, Box, Avatar, Typography, Card, CardContent, Button, CardActions } from '@mui/material';
+import { BallotOutlined } from '@mui/icons-material';
 import { SetButton } from '../Sets/component';
 
 // import { calculateNewTrainingMaxes, calculateUpdatedLifts, constructCompletedWorkout, getWorkoutForDay, validateWorkoutCompleted, workout } from '../util/workoutUtil';
 import { useEffect } from 'react';
+import { groupBy } from 'lodash';
 import { useSelector } from 'react-redux';
 import { titleCaseString } from '../util/textUtil';
 import { useAppDispatch, useAppSelector } from '../hooks';
@@ -12,6 +14,10 @@ import { workout } from '../util/workoutUtil';
 import * as profileActions from '../Profile/actions';
 import { Dispatch } from 'react';
 import { ObjectEntries } from '../util/util';
+import { useSearchParams } from 'react-router-dom';
+import { selectSetsForDay } from '../Sets/selectors';
+import { selectProgram } from '../Programs/selectors';
+import { SetDb } from '../Sets/types';
 
 const styles = {
   page: {
@@ -94,131 +100,136 @@ const styles = {
   }
 }
 
+const NeedToRegisterForProgram = () => {
+  return (
+    <Container component="main" maxWidth="xs">
+      <CssBaseline />
+      <Box
+        sx={{
+          marginTop: 8,
+          display: 'flex',
+          flexDirection: 'column',
+          alignItems: 'center',
+        }}
+      >
+        <Avatar sx={{ m: 1, bgcolor: 'secondary.main' }}>
+          <BallotOutlined />
+        </Avatar>
+        <Typography component="h1" variant="h5">
+          Workout
+        </Typography>
+        <Box sx={{ mt: 1 }}>
+          You need to register for a program before you can view this page.
+        </Box>
+      </Box>
+    </Container>
+  );
+};
+
+const NeedToSignIn = () => {
+  return (
+    <Container component="main" maxWidth="xs">
+      <CssBaseline />
+      <Box
+        sx={{
+          marginTop: 8,
+          display: 'flex',
+          flexDirection: 'column',
+          alignItems: 'center',
+        }}
+      >
+        <Avatar sx={{ m: 1, bgcolor: 'secondary.main' }}>
+          <BallotOutlined />
+        </Avatar>
+        <Typography component="h1" variant="h5">
+          Workout
+        </Typography>
+        <Box sx={{ mt: 1 }}>
+          You need to sign in or sign up.
+        </Box>
+      </Box>
+    </Container>
+  );
+};
+
 export const Workout = () => {
   const myProfile = useSelector(selectMyProfile);
-  console.log(myProfile);
+  const [searchParams, setSearchParams] = useSearchParams();
+  const week: number | null = (searchParams.get("week") ? Number(searchParams.get("week")) : null) ?? myProfile?.week ?? null;
+  const day: number | null = (searchParams.get("day") ? Number(searchParams.get("day")) : null) ?? myProfile?.day ?? null;
+  const programRegistrationId: string | null = searchParams.get("programRegistrationId") ?? myProfile?.programRegistrationId ?? null;
+  const sets = useSelector(selectSetsForDay({ week, day, programRegistrationId }));
+  console.log(sets);
+  const setsByMovement : { [key: string]: SetDb[] } = groupBy(sets, 'movement');
+  console.log(setsByMovement)
+  const movements : string[] = [];//setsByMovement.keys();
+  const profile = useSelector(selectMyProfile);
+  const program = useSelector(selectProgram(profile?.programId));
   const dispatch = useAppDispatch();
-  // const [week, setWeek]: [number, Dispatch<SetStateAction<number>>] = useState(myProfile ? myProfile.programSettings.week : 1);
-  // const [day, setDay]: [number, Dispatch<SetStateAction<number>>] = useState(myProfile ? myProfile.programSettings.day : 1);
-  const [currentWorkout, setCurrentWorkout]: [workout[] | undefined,  Function] = useState();
-  const [completedSets, setCompletedSets]: [{[key: string]: number[]}, Function] = useState({});
-  const [lastSetReps, setLastSetReps]: [{[key: string]: number}, Dispatch<SetStateAction<{[key: number]: number}>>] = useState({});
+  if (!programRegistrationId
+    || !week
+    || !day
+    || !program) {
+    return (<NeedToRegisterForProgram />);
+  }
 
-  // useEffect(() => {
-  //   if (myProfile) setCurrentWorkout(getWorkoutForDay(week, day, myProfile));
-  // }, [week, day, myProfile]);
+  if (!myProfile) {
+    return (<NeedToSignIn />);
+  }
 
-  // useEffect(() => {
-  //   if (currentWorkout) {
-  //     const defaultLastSets: {[key: string]: number} = {};
-  //     currentWorkout.forEach((workout: workout) => {
-  //       defaultLastSets[workout.name] = workout.sets[workout.sets.length - 1];
-  //     });
-
-  //     setLastSetReps(defaultLastSets);
-  //   }
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  // }, [currentWorkout]);
-
-  // useEffect(() => {
-  //   if (
-  //     myProfile?.completedWorkouts?.[week]?.[day]
-  //   ) {
-  //     setCurrentWorkout(myProfile.completedWorkouts[week][day]);
-  //     const completedWorkout: {[key: string]: number[]} = {};
-  //     ObjectEntries(myProfile.completedWorkouts[week][day]).forEach(([_, workout]: [string, workout]) => {
-  //       completedWorkout[workout.name] = Array.from(Array(workout.sets.length).keys());
-  //     });
-  //     console.log(completedWorkout);
-  //     setCompletedSets(completedWorkout);
-  //   }
-  // }, [week, day, myProfile]);
-
-  // const addRep = (workout: workout) => {
-  //   if (lastSetReps[workout.name]) {
-  //     setLastSetReps({...lastSetReps, [workout.name]: lastSetReps[workout.name] + 1});
-  //   } else {
-  //     setLastSetReps({...lastSetReps, [workout.name]: workout.sets[workout.sets.length - 1] + 1});
-  //   }
-  // };
-
-  // const minusRep = (workout: workout) => {
-  //   if (lastSetReps[workout.name]) {
-  //     setLastSetReps({...lastSetReps, [workout.name]: lastSetReps[workout.name] - 1});
-  //   } else {
-  //     setLastSetReps({...lastSetReps, [workout.name]: workout.sets[workout.sets.length - 1] - 1});
-  //   }
-  // };
-
-  // const completeWorkout = () => {
-  //   const workoutCompleted: boolean = validateWorkoutCompleted(completedSets, currentWorkout);
-  //   if (!workoutCompleted || !currentWorkout) {
-  //   } else {
-  //     const completedWorkout = constructCompletedWorkout(week, day, currentWorkout, lastSetReps);
-  //     const newTrainingMaxes = calculateNewTrainingMaxes(myProfile, currentWorkout, lastSetReps);
-  //     const updatedLifts = calculateUpdatedLifts(currentWorkout, lastSetReps);
-  //     dispatch(profileActions.completeWorkout(completedWorkout, newTrainingMaxes, updatedLifts));
-  //   }
-  // }
-
-  if (!myProfile) return <div>Error, no current user</div>;
-  
-  return <div>Page under construction</div>;
-
-  // return (
-    // <Grid container sx={styles.page}>
-    //   <Grid container sx={styles.contentContainer}>
-    //     <Grid container sx={styles.workoutContainerTitle}>
-    //       <Typography sx={styles.workoutTitle}>Workout</Typography>
-    //       <Grid item sx={styles.dropdownContainer}>
-    //         <FormControl variant="standard" sx={styles.dayDropdown}>
-    //           <InputLabel id="week-select-label">Week</InputLabel>
-    //           <Select
-    //             labelId="week-select-label"
-    //             id="week-select"
-    //             value={week}
-    //             label="Week"
-    //             onChange={(event) => setWeek(+event.target.value)}
-    //           >
-    //             {Object.keys(myProfile.program.setScheme.weeks).map((weekNumber: string) => (
-    //               <MenuItem key={weekNumber} value={weekNumber}>{weekNumber}</MenuItem>
-    //             ))}
-    //           </Select>
-    //         </FormControl>
-    //         <FormControl variant="standard" sx={styles.dayDropdown}>
-    //           <InputLabel id="day-select-label">Day</InputLabel>
-    //           <Select
-    //             labelId="day-select-label"
-    //             id="day-select"
-    //             value={day}
-    //             label="Day"
-    //             onChange={(event) => setDay(+event.target.value)}
-    //           >
-    //             {Array.from(Array(myProfile.programSettings.daysPerWeek).keys())
-    //               .map((val: number) => (
-    //                 <MenuItem key={val} value={val + 1}>{val + 1}</MenuItem>
-    //               )
-    //             )}
-    //           </Select>
-    //         </FormControl>
-    //       </Grid>
-    //     </Grid>
-    //     {currentWorkout && currentWorkout.map((workout: workout, workoutIndex: number) => (
-    //       <Grid item container sx={styles.movementsContainer} key={workout.name}>
-    //         <Grid item>
-    //           <Typography sx={styles.movementName}>{titleCaseString(workout.name)}</Typography>
-    //           <Typography sx={styles.movementWeight}>{workout.weight + ' lbs'}</Typography>
-    //         </Grid>
-    //         <Grid item sx={styles.setContainer}>
-    //           { workout.sets.map((set: number, index: number) => ( <SetButton setId={'asd'} /> ) ) }
-    //           <Grid item key={`${workout.name}-add`} sx={styles.setRepButton} onClick={() => addRep(workout)}>+</Grid>
-    //           <Grid item key={`${workout.name}-minus`} sx={styles.setRepButton} onClick={() => minusRep(workout)}>-</Grid>
-    //         </Grid>
-    //       </Grid>
-    //       )
-    //     )}
-    //     <button onClick={completeWorkout}>complete workout placeholder button</button>
-    //   </Grid>
-    // </Grid>
-  // )
+  return (
+    <Grid container sx={styles.page}>
+      <Grid container sx={styles.contentContainer}>
+        <Grid container sx={styles.workoutContainerTitle}>
+          <Typography sx={styles.workoutTitle}>Workout</Typography>
+          <Grid item sx={styles.dropdownContainer}>
+            <FormControl variant="standard" sx={styles.dayDropdown}>
+              <InputLabel id="week-select-label">Week</InputLabel>
+              <Select
+                labelId="week-select-label"
+                id="week-select"
+                value={week}
+                label="Week"
+              // onChange={(event) => setWeek(+event.target.value)}
+              >
+                {Object.keys(program.setScheme.weeks).map((weekNumber: string) => (
+                  <MenuItem key={weekNumber} value={weekNumber}>{weekNumber}</MenuItem>
+                ))}
+              </Select>
+            </FormControl>
+            <FormControl variant="standard" sx={styles.dayDropdown}>
+              <InputLabel id="day-select-label">Day</InputLabel>
+              <Select
+                labelId="day-select-label"
+                id="day-select"
+                value={day}
+                label="Day"
+              // onChange={(event) => setDay(+event.target.value)}
+              >
+                {Array.from(Array(program.daysPerWeek).keys())
+                  .map((val: number) => (
+                    <MenuItem key={val} value={val + 1}>{val + 1}</MenuItem>
+                  )
+                  )}
+              </Select>
+            </FormControl>
+          </Grid>
+        </Grid>
+        {movements.map((movement: string) => {
+          // <Grid item container sx={styles.movementsContainer} key={movement}>
+          //   <Grid item>
+          //     <Typography sx={styles.movementName}>{titleCaseString(movement)}</Typography>
+          //     <Typography sx={styles.movementWeight}>{workout.weight + ' lbs'}</Typography>
+          //   </Grid>
+          //   <Grid item sx={styles.setContainer}>
+          //     {sets.map((set: number, index: number) => (<SetButton setId={set.setId} />))}
+          //     <Grid item key={`${set.setId}-add`} sx={styles.setRepButton} onClick={() => addRep(workout)}>+</Grid>
+          //     <Grid item key={`${set.setId}-minus`} sx={styles.setRepButton} onClick={() => minusRep(workout)}>-</Grid>
+          //   </Grid>
+          // </Grid>
+        })}
+        {/* <button onClick={completeWorkout}>complete workout placeholder button</button> */}
+      </Grid>
+    </Grid>
+  )
 };
